Fix product filter script
On brand pages, there were only products of that brand. Now on all brand pages and the shop page, all products get loaded. When a brand page opens, the productfilter is initialized once. When user clicks on a brand or gender filter, it uses a tweaked version of the product filter function because it needs the  object.